BibleGateway: Fix bitrotted scraper, use mobile.*
Biblegateway has changed a fair bit. On the one hand it now claims to be
XHTML so it is more valid than it was. On the other hand the interface
to different bible versions has changed and there are now about 3
different ways of referring to a bible version. Some use numeric ids
still, but most use the full name, munged into the url. All use a short
name for the searching.

We use the mobile version at mobile.biblegateway.com as it's more likely
to be valid (mobile browsers are more pedantic), and likely to be more
compact (to save bandwidth).

The list of versions is now extracted directly from the versions list at
http://mobile.biblegateway.com/versions/. The table is read and the urls
to the version info pages are extracted.

The version info pages then provide more info, as well as a table
linking to every chapter in the Bible, sorted into rows for each book,
which can be read as they were before.

When reading the actual bible text, use the contents of sup tags instead
of the value attribute, and clip the footnotes section.